Switching to 2 clients and 1 server
My testing showed that after getting rid of redundancy, the bottleneck
in this test was the client, not the server side, so the test was not
sensitive to server side performance changes, which is what I'm trying
to test.

Increasing the load on the server to see if the test will measure an
impact.